Move TransactionBuilder out of rpc so it can be used in wasm
