+++ rust/psibase/src/lib.rs
@@ -33,6 +33,7 @@ mod time;
 mod to_bin;
 mod to_key;
 mod trace;
+mod transaction_builder;
 mod web_services;
 
 pub use account_number::*;
@@ -65,6 +66,7 @@ pub use time::*;
 pub use to_bin::*;
 pub use to_key::*;
 pub use trace::*;
+pub use transaction_builder::*;
 pub use web_services::*;
 
 use internal_macros::*;
